+++ src/lib/compiler/scriptBuilder.ts
@@ -1876,6 +1876,84 @@ class ScriptBuilder {
     this._addCmd("VM_ACTOR_MOVE_TO", addr);
   };
 
+  _actorMoveToInit = (addr: string, attr: string) => {
+    this._addCmd("VM_ACTOR_MOVE_TO_INIT", addr, attr);
+  };
+
+  _actorMoveToX = (addr: string, attr: string) => {
+    this._addCmd("VM_ACTOR_MOVE_TO_X", addr, attr);
+  };
+
+  _actorMoveToY = (addr: string, attr: string) => {
+    this._addCmd("VM_ACTOR_MOVE_TO_Y", addr, attr);
+  };
+
+  _actorMoveToXY = (addr: string, attr: string) => {
+    this._addCmd("VM_ACTOR_MOVE_TO_XY", addr, attr);
+  };
+
+  _actorMoveToSetDirX = (addr: string) => {
+    this._addCmd("VM_ACTOR_MOVE_TO_SET_DIR_X", addr);
+  };
+
+  _actorMoveToSetDirY = (addr: string) => {
+    this._addCmd("VM_ACTOR_MOVE_TO_SET_DIR_Y", addr);
+  };
+
+  _actorMoveToOps = (
+    addr: string,
+    attr: string,
+    moveType: ScriptBuilderMoveType,
+    moveX: boolean,
+    moveY: boolean,
+    lockDirection: ScriptBuilderAxis[],
+  ) => {
+    const lockDirX = lockDirection.includes("x");
+    const lockDirY = lockDirection.includes("y");
+
+    if (!moveX && !moveY) {
+      return;
+    }
+
+    this._actorMoveToInit(addr, attr);
+    if (moveX && !moveY) {
+      if (!lockDirX) {
+        this._actorMoveToSetDirX(addr);
+      }
+      this._actorMoveToX(addr, attr);
+    } else if (moveY && !moveX) {
+      if (!lockDirY) {
+        this._actorMoveToSetDirY(addr);
+      }
+      this._actorMoveToY(addr, attr);
+    } else if (moveType === "horizontal") {
+      if (!lockDirX) {
+        this._actorMoveToSetDirX(addr);
+      }
+      this._actorMoveToX(addr, attr);
+      if (!lockDirY) {
+        this._actorMoveToSetDirY(addr);
+      }
+      this._actorMoveToY(addr, attr);
+    } else if (moveType === "vertical") {
+      if (!lockDirY) {
+        this._actorMoveToSetDirY(addr);
+      }
+      this._actorMoveToY(addr, attr);
+      if (!lockDirX) {
+        this._actorMoveToSetDirX(addr);
+      }
+      this._actorMoveToX(addr, attr);
+    } else if (moveType === "diagonal") {
+      if (!lockDirY) {
+        this._actorMoveToSetDirY(addr);
+      } else if (!lockDirX) {
+        this._actorMoveToSetDirX(addr);
+      }
+      this._actorMoveToXY(addr, attr);
+    }
+  };
+
   _actorMoveCancel = (addr: string) => {
     this._addCmd("VM_ACTOR_MOVE_CANCEL", addr);
   };
@@ -3013,7 +3091,7 @@ extern void __mute_mask_${symbol};
     x: number,
     y: number,
     useCollisions: boolean,
-    moveType: ScriptBuilderMoveType,
+    moveType: ScriptBuilderMoveType = "horizontal",
     units: DistanceUnitType = "tiles",
   ) => {
     const actorRef = this._declareLocal("actor", 4);
@@ -3034,7 +3112,7 @@ extern void __mute_mask_${symbol};
     variableX: string,
     variableY: string,
     useCollisions: boolean,
-    moveType: ScriptBuilderMoveType,
+    moveType: ScriptBuilderMoveType = "horizontal",
     units: DistanceUnitType = "tiles",
   ) => {
     const actorRef = this._declareLocal("actor", 4);
@@ -3066,8 +3144,9 @@ extern void __mute_mask_${symbol};
     valueX: ScriptValue,
     valueY: ScriptValue,
     collideWith: boolean | Array<"walls" | "actors">,
-    moveType: ScriptBuilderMoveType,
+    moveType: ScriptBuilderMoveType = "horizontal",
     units: DistanceUnitType = "tiles",
+    lockDirection: ScriptBuilderAxis[] = [],
   ) => {
     const actorRef = this._declareLocal("actor", 4);
     const stackPtr = this.stackPtr;
@@ -3079,6 +3158,8 @@ extern void __mute_mask_${symbol};
     const optimisedY = optimiseScriptValue(
       scriptValueToSubpixels(valueY, units),
     );
+
+    const attr = toASMMoveFlags(moveType, collideWith);
 
     const [rpnOpsX, fetchOpsX] = precompileScriptValue(optimisedX, "x");
     const [rpnOpsY, fetchOpsY] = precompileScriptValue(optimisedY, "y");
@@ -3100,13 +3181,10 @@ extern void __mute_mask_${symbol};
     this._performValueRPN(rpn, rpnOpsY, localsLookup);
     rpn.refSet(this._localRef(actorRef, 2));
 
-    rpn.int16(toASMMoveFlags(moveType, collideWith));
-    rpn.refSet(this._localRef(actorRef, 3));
-
     rpn.stop();
     this._addComment(`-- Move Actor`);
     this.actorSetById(actorId);
-    this._actorMoveTo(actorRef);
+    this._actorMoveToOps(actorRef, attr, moveType, true, true, lockDirection);
     this._assertStackNeutral(stackPtr);
     this._addNL();
   };
@@ -3115,7 +3193,7 @@ extern void __mute_mask_${symbol};
     x = 0,
     y = 0,
     useCollisions = false,
-    moveType: ScriptBuilderMoveType,
+    moveType: ScriptBuilderMoveType = "horizontal",
     units: DistanceUnitType = "tiles",
   ) => {
     const actorRef = this._declareLocal("actor", 4);
@@ -3151,8 +3229,9 @@ extern void __mute_mask_${symbol};
     valueX: ScriptValue,
     valueY: ScriptValue,
     collideWith: boolean | Array<"walls" | "actors">,
-    moveType: ScriptBuilderMoveType,
+    moveType: ScriptBuilderMoveType = "horizontal",
     units: DistanceUnitType = "tiles",
+    lockDirection: ScriptBuilderAxis[] = [],
   ) => {
     const stackPtr = this.stackPtr;
     this._addComment("Actor Move Relative");
@@ -3166,6 +3245,8 @@ extern void __mute_mask_${symbol};
     if (!moveX && !moveY) {
       return;
     }
+
+    const attr = toASMMoveFlags(moveType, collideWith, true, units);
 
     const [rpnOpsX, fetchOpsX] = precompileScriptValue(
       optimiseScriptValue(scriptValueToSubpixels(valueX, units)),
@@ -3188,12 +3269,11 @@ extern void __mute_mask_${symbol};
     rpn.actorId(actorId); // Actor ID
     this._performValueRPN(rpn, rpnOpsX, localsLookup2); // X Value
     this._performValueRPN(rpn, rpnOpsY, localsLookup2); // Y Value
-    rpn.int16(toASMMoveFlags(moveType, collideWith, true, units)); // Move Flags
     rpn.stop();
 
     this._addComment(`-- Move Actor`);
-    this._actorMoveTo(".ARG3");
-    this._stackPop(4);
+    this._actorMoveToOps(".ARG2", attr, moveType, moveX, moveY, lockDirection);
+    this._stackPop(3);
     this._assertStackNeutral(stackPtr);
     this._addNL();
   };